requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
Flask==3.1.3
google-genai==1.46.0
gunicorn==25.1.0
idna==3.11
pydantic==2.10.5
PyJWT==2.11.0
pytz==2026.1.post1
requests==2.32.5
urllib3==2.6.3
